MAINT: Move warning
Having time dimension chunked is actually often optimal.
However, when bootstrapping percentiles it is almost
necessary because, of the use of map_block within bootstapping.
